Display user role names correctly for synced users
