+++ graylog2-web-interface/src/components/authentication/BackendDetails/BackendDetails.jsx
@@ -1,6 +1,9 @@
 // @flow strict
 import * as React from 'react';
+import { useEffect, useState } from 'react';
 
+import { Spinner } from 'components/common';
+import AuthzRolesDomain from 'domainActions/roles/AuthzRolesDomain';
 import { getAuthServicePlugin } from 'util/AuthenticationService';
 import AuthenticationBackend from 'logic/authentication/AuthenticationBackend';
 import SectionGrid from 'components/common/Section/SectionGrid';
@@ -12,10 +15,21 @@ type Props = {
 };
 
 const BackendDetails = ({ authenticationBackend }: Props) => {
+  const [{ list: roles }, setPaginatedRoles] = useState({ list: undefined });
   const authService = getAuthServicePlugin(authenticationBackend.config.type);
+
+  useEffect(() => {
+    const getUnlimited = [1, 0, ''];
+
+    AuthzRolesDomain.loadRolesPaginated(...getUnlimited).then((newPaginatedRoles) => newPaginatedRoles && setPaginatedRoles(newPaginatedRoles));
+  }, []);
 
   if (!authService) {
     return `No authentication service plugin configured for type "${authenticationBackend.config.type}"`;
+  }
+
+  if (!roles) {
+    return <Spinner />;
   }
 
   const { configDetailsComponent: BackendConfigDetails } = authService;
@@ -23,7 +37,7 @@ const BackendDetails = ({ authenticationBackend }: Props) => {
   return (
     <SectionGrid>
       <div>
-        <BackendConfigDetails authenticationBackend={authenticationBackend} />
+        <BackendConfigDetails authenticationBackend={authenticationBackend} roles={roles} />
       </div>
       <div>
         <SectionComponent title="Synchronized Users">
