+++ graylog2-web-interface/src/components/authentication/BackendDetailsActive/BackendDetailsActive.jsx
@@ -1,9 +1,12 @@
 // @flow strict
 import * as React from 'react';
+import { useEffect, useState } from 'react';
 
 import { getAuthServicePlugin } from 'util/AuthenticationService';
 import AuthenticationBackend from 'logic/authentication/AuthenticationBackend';
 import SectionGrid from 'components/common/Section/SectionGrid';
+import { Spinner } from 'components/common';
+import AuthzRolesDomain from 'domainActions/roles/AuthzRolesDomain';
 
 import SyncedUsersSection from './SyncedUsersSection';
 import SyncedTeamsSection from './SyncedTeamsSection';
@@ -13,10 +16,21 @@ type Props = {
 };
 
 const BackendDetailsActive = ({ authenticationBackend }: Props) => {
+  const [{ list: roles }, setPaginatedRoles] = useState({ list: undefined });
   const authService = getAuthServicePlugin(authenticationBackend.config.type);
+
+  useEffect(() => {
+    const getUnlimited = [1, 0, ''];
+
+    AuthzRolesDomain.loadRolesPaginated(...getUnlimited).then((newPaginatedRoles) => newPaginatedRoles && setPaginatedRoles(newPaginatedRoles));
+  }, []);
 
   if (!authService) {
     return `No authentication service plugin configured for type "${authenticationBackend.config.type}"`;
+  }
+
+  if (!roles) {
+    return <Spinner />;
   }
 
   const { configDetailsComponent: BackendConfigDetails } = authService;
@@ -24,11 +38,11 @@ const BackendDetailsActive = ({ authenticationBackend }: Props) => {
   return (
     <SectionGrid>
       <div>
-        <BackendConfigDetails authenticationBackend={authenticationBackend} />
+        <BackendConfigDetails authenticationBackend={authenticationBackend} roles={roles} />
       </div>
       <div>
-        <SyncedUsersSection />
-        <SyncedTeamsSection />
+        <SyncedUsersSection roles={roles} />
+        <SyncedTeamsSection roles={roles} />
       </div>
     </SectionGrid>
   );
