+++ graylog2-web-interface/src/components/authentication/BackendDetailsActive/SyncedUsersSection/SyncedUsersOverviewItem.jsx
@@ -4,6 +4,7 @@ import styled from 'styled-components';
 import { Link } from 'react-router';
 import { LinkContainer } from 'react-router-bootstrap';
 
+import Role from 'logic/roles/Role';
 import Routes from 'routing/Routes';
 import AuthenticationDomain from 'domainActions/authentication/AuthenticationDomain';
 import UserOverview from 'logic/users/UserOverview';
@@ -12,6 +13,7 @@ import RolesCell from 'components/users/UsersOverview/UserOverviewItem/RolesCell
 
 type Props = {
   user: UserOverview,
+  roles: Immutable.List<Role>,
 };
 
 const ActtionsWrapper = styled(ButtonToolbar)`
@@ -24,10 +26,15 @@ const SyncedUsersOverviewItem = ({
     enabled,
     fullName,
     id,
-    roles,
+    roles: userRolesIds,
     username,
   },
+  roles,
 }: Props) => {
+  const userRolesNames = userRolesIds.map((roleId) => {
+    return roles.find((role) => role.id === roleId)?.name;
+  });
+
   return (
     <tr key={username}>
       <td className="limited">
@@ -36,7 +43,7 @@ const SyncedUsersOverviewItem = ({
         </Link>
       </td>
       <td className="limited">{fullName}</td>
-      <RolesCell roles={roles} />
+      <RolesCell roles={userRolesNames} />
       <td className="limited">
         <ActtionsWrapper>
           {enabled
